Move layouts for the tab sections to a new package 'tab_sections'.
